use string type for VALUEn columns

+++ pdp/data/noaa/ghcnd.py
@@ -46,7 +46,7 @@ class GHCNDParseTextFiles(SparkTask):
 
         def daily_observation(day: int) -> list[Column]:
             start_width = total_id_width + (day - 1) * observation_total_width
-            value = TEXT_COL.substr(start_width, VALUE_CHARS).cast("int").alias(f"VALUE{day}")
+            value = TEXT_COL.substr(start_width, VALUE_CHARS).alias(f"VALUE{day}")
             flags = [
                 TEXT_COL.substr(start_width + VALUE_CHARS + i, 1).alias(f"{f}FLAG{day}")
                 for i, f in enumerate(SINGLE_CHAR_FLAGS)
